+++ zds/tutorialv2/utils.py
@@ -1,5 +1,4 @@
 # coding: utf-8
-import os
 import shutil
 from collections import OrderedDict
 from datetime import datetime
@@ -7,6 +6,7 @@ from urllib import urlretrieve
 from urlparse import urlparse
 
 import cairosvg
+import os
 from PIL import Image as ImagePIL
 from django.http import Http404
 from django.utils.translation import ugettext_lazy as _
@@ -16,6 +16,7 @@ from uuslug import slugify
 
 from zds import settings
 from zds.member.models import User
+from zds.notification import signals
 from zds.tutorialv2 import REPLACE_IMAGE_PATTERN, VALID_SLUG
 from zds.utils import get_current_user
 from zds.utils import slugify as old_slugify
@@ -164,6 +165,7 @@ def mark_read(content, user=None):
                 content=content,
                 user=user)
             a.save()
+            signals.content_read.send(sender=content.__class__, instance=content, user=user)
 
 
 class TooDeepContainerError(ValueError):
